Cambio de tipo de datos para el tema de la version (tam. fijo)

+++ trunk/Ejercicio 6/v5 dist. varios brokers con anillo y lider/Anillo/sender.c
@@ -17,8 +17,8 @@ main(int argc, char *argv[])
     char mostrar[300];
     
     key_t key = ftok("/tmp/buchwaldipcs", SHM_VERSION);
-    int shmversion = shmget(key, sizeof(int) , 0660 | IPC_CREAT);
-    int* id = (int*)shmat(shmversion, NULL, 0);   
+    int shmversion = shmget(key, sizeof(uint64_t) , 0660 | IPC_CREAT);
+    uint64_t* id = (uint64_t*)shmat(shmversion, NULL, 0);   
     
     sprintf(mostrar,"[MASTER] --> ID: %d\t| pid: %d\n",*id, getpid());
     write(fileno(stdout),mostrar,strlen(mostrar));
